Correct NMI handling to better reflect how it works
I misunderstood how NMI was "edge-triggered" and assumed that the PPU
just somehow "pulsed" the NMI line. I now realize that it's actually to
do with the fact that the 6502 only performs NMI interrupt when NMI is
going from LOW to HIGH, unlike IRQ, which interrupts as long as IRQ is
high. This change has no real effect, but it helps with actually
understanding how it works.